+++ src/main/java/com/sparta/todayeats/address/service/AddressService.java
@@ -7,8 +7,8 @@ import com.sparta.todayeats.address.entity.Address;
 import com.sparta.todayeats.address.repository.AddressRepository;
 import com.sparta.todayeats.global.exception.*;
 import com.sparta.todayeats.global.service.UserAuthorizationService;
-import com.sparta.todayeats.user.domain.entity.User;
-import com.sparta.todayeats.user.domain.repository.UserRepository;
+import com.sparta.todayeats.user.entity.User;
+import com.sparta.todayeats.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
