+++ src/test/java/com/sparta/todayeats/address/service/AddressServiceTest.java
@@ -8,8 +8,8 @@ import com.sparta.todayeats.address.repository.AddressRepository;
 import com.sparta.todayeats.global.exception.AuthErrorCode;
 import com.sparta.todayeats.global.exception.BaseException;
 import com.sparta.todayeats.global.service.UserAuthorizationService;
-import com.sparta.todayeats.user.domain.entity.User;
-import com.sparta.todayeats.user.domain.repository.UserRepository;
+import com.sparta.todayeats.user.entity.User;
+import com.sparta.todayeats.user.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
